collider

+++ project/Engine/Audio/AudioManager.cpp
@@ -116,10 +116,10 @@ uint32_t AudioManager::LoadAudio(const std::string& fileName) {
 	pMFMediaType->SetGUID(MF_MT_MAJOR_TYPE, MFMediaType_Audio);
 	pMFMediaType->SetGUID(MF_MT_SUBTYPE, MFAudioFormat_PCM);
 
-	pMFSourceReader->SetCurrentMediaType(MF_SOURCE_READER_FIRST_AUDIO_STREAM, nullptr, pMFMediaType);
+	pMFSourceReader->SetCurrentMediaType(DWORD(MF_SOURCE_READER_FIRST_AUDIO_STREAM), nullptr, pMFMediaType);
 	pMFMediaType->Release();
 	pMFMediaType = nullptr;
-	pMFSourceReader->GetCurrentMediaType(MF_SOURCE_READER_FIRST_AUDIO_STREAM, &pMFMediaType);
+	pMFSourceReader->GetCurrentMediaType(DWORD(MF_SOURCE_READER_FIRST_AUDIO_STREAM), &pMFMediaType);
 
 	WAVEFORMATEX* waveFormat{nullptr};
 	MFCreateWaveFormatExFromMFMediaType(pMFMediaType,&waveFormat,nullptr);
@@ -128,7 +128,7 @@ uint32_t AudioManager::LoadAudio(const std::string& fileName) {
 	while (1) {
 		IMFSample* pMFSample{nullptr};
 		DWORD dwStreamFlags{0};
-		pMFSourceReader->ReadSample(MF_SOURCE_READER_FIRST_AUDIO_STREAM,0,nullptr,&dwStreamFlags,nullptr,&pMFSample);
+		pMFSourceReader->ReadSample(DWORD(MF_SOURCE_READER_FIRST_AUDIO_STREAM),0,nullptr,&dwStreamFlags,nullptr,&pMFSample);
 	
 		if (dwStreamFlags & MF_SOURCE_READERF_ENDOFSTREAM) {
 			break;
